requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
Pillow==9.3.0
django-filter==21.1